Connect the user and auth nodes
The auth node is now using a hitting a real user node using requests
instead of pretending to hit one.

The nodes are both running in a Docker container and are able to
communicate because they are on the same Docker network (addb). At this
time the command `docker create network addb` needs to be run before
starting up the auth and user Docker containers.

The next step is to use Docker swarm to hook up everything so that we can
improve this piecemeal build process.