Modify User service Docker setup to connect to PG
We've pulled a Postgres Docker container onto our machine and connected
to it by adding it to the Docker network. We've even mounted the
configuration file directory onto the user service node so we can make
configuration changes on the fly.

This change required installing the pip module psycopg2 and the Ubuntu
apt software python-psycopg2.

The containers are able to talk to each other. The user service can
write data to the Postgres database.

This interaction exposed a bug in the way we are writing passwords to
the User table. The encrypted password is being written in byte format
instead of string format. Postgres stores this data as a hexadecimal
string starting with \x. The encrypted password get re-encrypted because
the UserDataObject sees it as a "string" and it bcrypt is no longer
able to validate the password correctly.

We are going to fix this bug next.