Modify password storage to work with a real db
When we connected the application PostgreSQL we hit issues with password
validation because we were dealing with byte-encoding of passwords in a
haphazard way.

We've changed how we deal with passwords. We will now store both the
password and its hashed variant in string format in the UserDataObject.
Passwords will remain in string format unless they are being validated
by bcrypt, at which point they will be converted to bytes on-the-fly.

We've also moved the password validation logic into the UserDataObject.

By ensuring that passwords are in a consistent format and putting all of
the password validation logic in one place we have increased the
maintainability of the application. Also, now that passwords are always
going to be treated as strings PostgreSQL will have no issues with
storing the information.