docker-composed the services that we have so far
Before we had to start up:
    * User
    * Authentication
    * PostgreSQL

one-by-one either through a shell script or a copy-pasted shell command.

We've now introduced docker-compose to our project. With this new
docker-compose.yml file all we need to do is run `docker-compose up` in
the project's root directory which will bring up all three components of
the system at once. The components will also be able to communicate with
one another because of the default network built into docker-compose.

With this setup we will now be able to introduce new services with relative ease.