Add a test runner to authentication and user
We've installed the Python testrunner `nose` which we are now running
via shellscript in both the authentication and user services. This will
simplify testing as we can just run the test.sh script instead of going
into the test directory and running each test file one-by-one.

In running my tests I stumbled upon some breaking tests which I took the
liberty of fixing.