Initialize contact router & initial db connection
We're on the way towards implementing an endpoint which will let us add
a new contact to the system for a given user.

We need to create schemas so we can store users and contacts. From there
the endpoint can be connected to the database stuff through some kind of
controller.