Add table initialization for user and contact
We are now initializing the core tables that are going to be used by the
contact service to store information. We will expand on these table
schemas over time but for now, this will get us started.

The next thing we need to do is wrap queries against these tables with
contact and user models.