Add method to fetch users from database
We now have a static method on the user model that we can use to fetch a
user from the database given their uuid.

I discovered a bug while I was coding this. For some reason I thought
that default function arguments in javascript work the same way that
they do in python and this false assumption casued some issues when I
was trying to initialize a user object given an id and created_at
timestamp.