Switch from console.log to winston for logging
Logs will now be json strings that can be consumed by another system in
the distant future. For now we will have lots of cool information like
log levels and timestamps which will be helpful for debugging purposes.

For now we're logging to the console but eventually we'll spit logs into
files too.