Containerize postgres for contact app
The contact app needs a postgres instance running so we've codified it
into the docker-compose file that we already have.

We had to create a new image for the postgres and as a part of that
we've created a sql script to initialize the dev, test, and prod
databases that will be needed by the application.

It's easier to have docker handle it for us instead of trying to do it
at the application level, though if we find that we need the level of
control we can always change direction.