Move user service config into repository
Having this information live outside of the repository seemed like a good idea at first, but after some more development work, that is no longer the case.

I don't want to have to remember to create this configuration file if I clone this application on a different machine. Most likely, I'm going to forget anyway. Also, I'd have reverse engineer what the file is supposed to look like, which is a pain.