User can now register from UI
We can now take the user-supplied data from the registration form and pass it along to user node and ask it to register the user.

I added axios to the frontend node so that I could easily send a post request to the user node.

I had to modify the user node and install sanic-cors. Both the frontend and user nodes are going to live on the same network and without CORS they won't be able to communicate.

Now that users can register successfully, the next thing we need to do is grab a JWT from the authentication node.