Integrate frontend with auth node
We are not reaching out to the auth node and request a JWT for a newly created user.

In order to make this work we had to enable CORS in the auth node because everything is running localhost.

Now that we have both the uuid and token for a newly register user, we need to store this information into Vuex on the frontend so that we can refer to it for communication with the contact node.