Redesign contact card
I moved the "View" button to the right of the card and added some template data for email and phone number.

Now that we have some idea of what the UI for this is going to look like, we can move the contact card logic into a separate component that can then be fed props for contact data.