Create a Contact component
I moved the code for the Contact card into its own component. Now, when we finally fetch contact data from the backend, we can simply feed it into a series of Contact components.

The Contact components will eventually have business logic to handle some scenarios like edits and deletes, but for now we don't need any of that.