Add contact data fetch from frontend
We are now asking the contact service for contact data for the user that's currently logged in. This data is presented on the homepage of the application.

We were able to accomplish this easily using the built-in proxy server capabilities of Vue. Unfortunately I learned this **after** I already implemented CORS in the User and Auth services. Eventually I will go back and remove CORS from those services and then proceed to add them to the `vue.config.js` file.