Add a login component
I took the Register view and converted it into a register component. I then wrapped the register component with a Portal view (i.e. the portal to the application).

I created a new login component and now portal will toggle between them. I think that this flow is better than having to go to a new URL to log into the application. Also, now the entrance to the application lives in one place.