Add a login page
I've created a login page for the application. It consumes & validates a (username, password) pair, and if everything is good, it redirects the user to the main page of the application.

As a part of this enhancement I had to start proxying all of the other exisiting backend services which I previously got working thanks to the addition of CORS. When I added devServer proxy logic for the contact service, I broke the existing URL logic that was being used to access the auth and user services. Adding the auth and user  serviecs to the devServer[proxy] setting fixed the issue.

I also refactored the connectors and pulled all of them into their own files. This is because I now have to use the auth and user connectors in two frontend services: the register service and the login service.