Fix registration bug
Registering with a user id that already exists in the system was crashing the frontend. This is because, while we were checking to make sure that our post requests returned successfully, we didn't check to see if there were any error messages in the response data.

We are now unpacking the data and checking to see if there are any errors within, similar to what we're already doing in the `service/login.js`.