Refactor UI
I made two changes here in order to make some improvements.

Firstly, I changed the default landing page so that now, instead of being shown the registration dialog, you are shown the login dialog. Registration is something that only needs to be done once. Login, however, is the only thing that the user is going to do afterwards. Because it's likely to be more frequent, it should be the default.

Secondly, I moved the toolbar into its own vue component. Just trying to be more modular...